Finishes BattleCalculator and changes floats by doubles

+++ src/main/java/calculation/Calculation.java
@@ -2,6 +2,6 @@ package calculation;
 
 public interface Calculation {
 
-  double calculate(float defeatedPercentage);
+  double calculate(double defeatedPercentage);
 
 }
