+++ src/main/java/calculation/HalfDownDefeatedCalculation.java
@@ -3,11 +3,11 @@ package calculation;
 public class HalfDownDefeatedCalculation implements Calculation {
 
   @Override
-  public double calculate(float defeatedPercentage) {
+  public double calculate(double defeatedPercentage) {
     return minimumZero(defeatedPercentage - 15);
   }
 
-  private double minimumZero(float number) {
+  private double minimumZero(double number) {
     return (number < 0) ? 0 : number;
   }
 }
