+++ src/main/java/calculation/TopHalfDefeatedCalculation.java
@@ -3,7 +3,7 @@ package calculation;
 public class TopHalfDefeatedCalculation implements Calculation {
 
   @Override
-  public double calculate(float defeatedPercentage) {
+  public double calculate(double defeatedPercentage) {
     return - defeatedPercentage / 2 + 60;
   }
 }
