+++ src/main/java/calculation/VictorCalculation.java
@@ -3,7 +3,7 @@ package calculation;
 public class VictorCalculation implements Calculation {
 
   @Override
-  public double calculate(float defeatedPercentage) {
+  public double calculate(double defeatedPercentage) {
     return -9 * Math.pow(defeatedPercentage, 2) / 1000 + 100;
   }
 }
